requirements.txt
tensorflow
magenta
RPi.GPIO
smbus